# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateEvidenceResponse": {
        "properties": {
          "heading_path": {
            "items": {
              "type": "string"
            },
            "title": "Heading Path",
            "type": "array"
          },
          "locator": {
            "additionalProperties": true,
            "title": "Locator",
            "type": "object"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "quote_hash": {
            "title": "Quote Hash",
            "type": "string"
          },
          "section_id": {
            "format": "uuid",
            "title": "Section Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "section_id",
          "title",
          "text",
          "heading_path",
          "page_number",
          "locator",
          "quote_hash"
        ],
        "title": "CandidateEvidenceResponse",
        "type": "object"
      },
      "CandidateListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandidateResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "CandidateListResponse",
        "type": "object"
      },
...
      "CandidateResponse": {
        "properties": {
          "atomicity": {
            "title": "Atomicity",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "title": "Conditions",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/CandidateEvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "exceptions": {
            "items": {
              "type": "string"
            },
            "title": "Exceptions",
            "type": "array"
          },
          "extraction_job_id": {
            "format": "uuid",
            "title": "Extraction Job Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "source_title": {
            "title": "Source Title",
            "type": "string"
          },
          "source_version_id": {
            "format": "uuid",
            "title": "Source Version Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "suggested_destination_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Destination Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "verification_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Reason"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "extraction_job_id",
          "source_version_id",
          "source_title",
          "title",
          "body",
          "tags",
          "suggested_destination_id",
          "atomicity",
          "confidence",
          "status",
          "version",
          "verification_reason",
          "rejection_reason",
          "reviewed_at",
          "conditions",
          "exceptions",
          "model",
          "prompt_version",
          "created_at",
          "evidence"
        ],
        "title": "CandidateResponse",
        "type": "object"
      },
...
      "CandidateStatus": {
        "enum": [
          "pending_review",
          "needs_verification",
          "accepted",
          "rejected"
        ],
        "title": "CandidateStatus",
        "type": "string"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ExtractionJobResponse": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "parse_artifact_id": {
            "format": "uuid",
            "title": "Parse Artifact Id",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "provider_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Request Id"
          },
          "source_version_id": {
            "format": "uuid",
            "title": "Source Version Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job_id",
          "source_version_id",
          "parse_artifact_id",
          "status",
          "provider",
          "model",
          "prompt_version",
          "input_tokens",
          "output_tokens",
          "latency_ms",
          "provider_request_id",
          "error_code",
          "error_message",
          "started_at",
          "completed_at",
          "created_at"
        ],
        "title": "ExtractionJobResponse",
        "type": "object"
      },
      "ExtractionRunListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExtractionRunResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ExtractionRunListResponse",
        "type": "object"
      },
      "ExtractionRunResponse": {
        "properties": {
          "extraction": {
            "$ref": "#/components/schemas/ExtractionJobResponse"
          },
          "job": {
            "$ref": "#/components/schemas/JobResponse"
          },
          "source_title": {
            "title": "Source Title",
            "type": "string"
          }
        },
        "required": [
          "job",
          "extraction",
          "source_title"
        ],
        "title": "ExtractionRunResponse",
        "type": "object"
      },
      "JobKind": {
        "enum": [
          "source_ingest",
          "source_parse",
          "source_extract",
          "source_index"
        ],
        "title": "JobKind",
        "type": "string"
      },
...
      "JobResponse": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "source_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Version Id"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "space_id",
          "source_version_id",
          "kind",
          "status",
          "progress",
          "attempt_count",
          "retryable",
          "error_code",
          "error_message",
          "started_at",
          "finished_at",
          "created_at",
          "updated_at"
        ],
        "title": "JobResponse",
        "type": "object"
      },
...
      "ScheduleExtractionResponse": {
        "properties": {
          "extraction": {
            "$ref": "#/components/schemas/ExtractionJobResponse"
          },
          "job": {
            "$ref": "#/components/schemas/JobResponse"
          }
        },
        "required": [
          "job",
          "extraction"
        ],
        "title": "ScheduleExtractionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "D5 immutable source ingestion, parsing, structured AI extraction, and human candidate review API for the personal knowledge workbench.",
    "title": "Zixu Knowledge Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge-spaces/{space_id}/candidates": {
      "get": {
        "operationId": "list_extraction_candidates",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CandidateStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Candidates",
        "tags": [
          "extraction"
        ]
      }
    },
...
    "/api/v1/knowledge-spaces/{space_id}/extractions": {
      "get": {
        "operationId": "list_source_extractions",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Extractions",
        "tags": [
          "extraction"
        ]
      }
    },
...
    "/api/v1/knowledge-spaces/{space_id}/sources/{source_id}/versions/{version_id}/extraction": {
      "get": {
        "operationId": "get_source_extraction",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExtractionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Extraction",
        "tags": [
          "extraction"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/sources/{source_id}/versions/{version_id}/extractions": {
      "post": {
        "operationId": "schedule_source_extraction",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleExtractionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Schedule Extraction",
        "tags": [
          "extraction"
        ]
      }
    },
...
  }
}